rename add bookmark to add bookmark category
